Eliminate flicker during scrolling
In the current CodeMirror codebase, when scrolling through a long document, there is
flicker due to the fact that the mover div is hosted in an actual large scroll area. If
the user scrolls quickly, a blank portion of the scroll area is drawn before the mover
is moved into its new position. This commit creates a separate scrollbar with its own
internal (blank) div to set its scrollHeight. The main mover div only moves up and down
as necessary to account for the pixel offset between the current virtual scroll position
and the actual top of the rendered area (showingFrom).